Web client is now functional.
It is not connected to the main logic of the program, but it can be used by importing. Logging in, starting a game, and playing moves (including a response showing whether the move was played or not) are implemented. Waiting for opponent move is still not.

Additionally the README was update, and some structure was added to the controller.