Change web_client.play_move()'s wait for move logic to wait for a change in player's turn instead of a change in last move.
For this to be reasonable the signature changed to include the moving player's color. Where we return False immediately if it's not the given color's player turn.
This is useful in general for the driver in case in the future
demonstration/exhibition games on a signle client will be possible, as it
will greatly help with keeping the games in sync.

In addition, a move will now return False immediately if the corresponding
value at game[dotsData][y][x] is not 0 (this includes the case it doesn't
exist at all), as this means that either the move is not on the board at
all or the space is already occupied. This should reduce the amount of time we have to wait the full WAIT_FOR_MOVE duration.