Reconnect is now available.
if you either give an ID of a game to connect to to the htp-client, or automatically connect into a game when pressing play, the record of the game will be passed back in the return value of start_game (An empty game record will still be returned when a new game does start).

This change revealed a bug in web_client.parse_server_coordinates using the wrong coordinate set when calculating maximum possible x value. This should now be fixed and regression tests have been added.